move util* to duck/util, change Sinkable string to SinkableDomain as per pr comments
