+++ apis/duck/util/util_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1alpha1
+package util
 
 import (
 	"fmt"
@@ -23,8 +23,8 @@ import (
 
 	"encoding/json"
 
+	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
@@ -36,9 +36,9 @@ func getValidReference() corev1.ObjectReference {
 }
 
 type StatusWithExtraFields struct {
-	Value        string        `json:"value",omitempty"`
-	Another      string        `json:"another",omitempty"`
-	Subscribable *Subscribable `json:"subscribable,omitempty"`
+	Value        string                     `json:"value",omitempty"`
+	Another      string                     `json:"another",omitempty"`
+	Subscribable *duckv1alpha1.Subscribable `json:"subscribable,omitempty"`
 }
 
 type StatusWithMissingSubscribable struct {
@@ -46,33 +46,11 @@ type StatusWithMissingSubscribable struct {
 	Another string
 }
 
-func getUnstructured() unstructured.Unstructured {
-	g := Generational{
-		metav1.TypeMeta{
-			Kind:       "test_kind",
-			APIVersion: "test_kind",
-		},
-		metav1.ObjectMeta{Name: "test_name"},
-		GenerationalSpec{1234},
-	}
-	raw, err := json.Marshal(g)
-	if err != nil {
-		panic("failed to marshal")
-	}
-	var m map[string]interface{}
-	if err := json.Unmarshal(raw, &m); err != nil {
-		fmt.Printf("Failed to unmarshal: %s", err)
-		panic("failed to unmarshal")
-	}
-	return unstructured.Unstructured{Object: m}
-
-}
-
 func TestFromUnstructuredSubscription(t *testing.T) {
 	tcs := []struct {
 		name      string
 		in        unstructured.Unstructured
-		want      SubscribableStatus
+		want      duckv1alpha1.SubscribableStatus
 		wantError error
 	}{{
 		name: "Works with valid status",
@@ -81,9 +59,9 @@ func TestFromUnstructuredSubscription(t *testing.T) {
 				"apiVersion": "test",
 				"kind":       "test_kind",
 				"name":       "test_name",
-				"status":     SubscribableStatus{&Subscribable{getValidReference()}},
+				"status":     duckv1alpha1.SubscribableStatus{&duckv1alpha1.Subscribable{getValidReference()}},
 			}},
-		want: SubscribableStatus{&Subscribable{corev1.ObjectReference{
+		want: duckv1alpha1.SubscribableStatus{&duckv1alpha1.Subscribable{corev1.ObjectReference{
 			Kind:       "Channel",
 			APIVersion: "channels.eventing.knative.dev/v1alpha1",
 			Name:       "mychannel"}}},
@@ -97,12 +75,12 @@ func TestFromUnstructuredSubscription(t *testing.T) {
 				"name":       "test_name",
 				"status":     StatusWithMissingSubscribable{"first", "second"},
 			}},
-		want:      SubscribableStatus{},
+		want:      duckv1alpha1.SubscribableStatus{},
 		wantError: nil,
 	}, {
 		name:      "empty unstructured",
 		in:        unstructured.Unstructured{},
-		want:      SubscribableStatus{},
+		want:      duckv1alpha1.SubscribableStatus{},
 		wantError: nil,
 	}}
 	for _, tc := range tcs {
@@ -112,7 +90,7 @@ func TestFromUnstructuredSubscription(t *testing.T) {
 		}
 		fmt.Printf("Marshalled : %s", string(raw))
 
-		got := Subscription{}
+		got := duckv1alpha1.Subscription{}
 		err = FromUnstructured(tc.in, &got)
 		if err != tc.wantError {
 			t.Errorf("Unexpected error for %q: %v", string(tc.name), err)
